Move client utility types to client/util.rs
